fix(reservation-app): 예약 생성 API 분리
- 자원 생성 시 방과 좌석이 필수인 점을 Endpoint 단계에서 명확하게 표현함
