+++ gateway/src/main/resources/application.yml
@@ -27,7 +27,7 @@ spring:
             - id: reservation-service-route
               uri: http://localhost:8082
               predicates:
-                - Path=/api/v1/reservation/**,/api/v1/rooms/**,/api/v1/waitlist/**
+                - Path=/api/v1/reservations/**,/api/v1/rooms/**,/api/v1/waitlist/**
               filters:
                 - RewritePath=/api/v1/?(?<segment>.*), /$\{segment}
