+++ reservation/reservation-application/src/main/java/com/seatliberator/seatliberator/reservation/book/infrastructure/web/request/ReservationCreateRequest.java
@@ -3,8 +3,6 @@ package com.seatliberator.seatliberator.reservation.book.infrastructure.web.requ
 import java.time.Instant;
 
 public record ReservationCreateRequest(
-        String roomId,
-        String seatId,
         Instant startAt,
         Instant endAt
 ) {
